Vim: Override Rg command
We want to search from the root of the current buffer not the current
working directory of the file first opened in vim.

Think of a scenario when you have opened a entrypoint file but jumped
into one of its dependencies. You want to Rg in the dependency library
not at the entrypoint.