Gitignore: Prune to Only Temp Files
Commits should be made deliberately so that if binaries, archives or
other files are seen the the diff, they should be ignored then.

It causes problems to ingore them at the root because some project check
binaries in to build docker containers.  They build fine locally but
never get to the build agent.